Reorganize distributions in the same file
