--- numpyro/distributions/expon.py
@@ -1,42 +1,0 @@
-# Source code modified from scipy.stats._continous_distns.py
-#
-# Copyright (c) 2001, 2002 Enthought, Inc.
-# All rights reserved.
-#
-# Copyright (c) 2003-2019 SciPy Developers.
-# All rights reserved.
-
-import jax.numpy as np
-import jax.random as random
-
-from numpyro.distributions.distribution import jax_continuous
-
-
-class expon_gen(jax_continuous):
-    def _rvs(self):
-        u = random.uniform(self._random_state, self._size)
-        return -np.log(u)
-
-    def _cdf(self, x):
-        return -np.expm1(-x)
-
-    def _ppf(self, q):
-        return -np.log1p(-q)
-
-    def _sf(self, x):
-        return np.exp(-x)
-
-    def _logsf(self, x):
-        return -x
-
-    def _isf(self, q):
-        return -np.log(q)
-
-    def _stats(self):
-        return 1.0, 1.0, 2.0, 6.0
-
-    def _entropy(self):
-        return 1.0
-
-
-expon = expon_gen(name='expon')
