+++ numpyro/examples/minipyro.py
@@ -1,13 +1,12 @@
 import argparse
 
 import jax.numpy as np
-from jax import random, lax
+from jax import lax, random
 from jax.experimental import optimizers
 from jax.random import PRNGKey
 
 import numpyro.distributions as dist
-
-from numpyro.handlers import sample, param
+from numpyro.handlers import param, sample
 from numpyro.svi import elbo, svi
