+++ setup.cfg
@@ -8,6 +8,7 @@ not_skip = __init__.py
 skip_glob = .ipynb_checkpoints
 known_first_party = numpyro, tests
 known_third_party = opt_einsum, six, torch, torchvision
+multi_line_output = 3
 
 [tool:pytest]
 filterwarnings = error
